requirements.txt
langchain
langchain-community
langchain-openai
requests
click
python-dotenv
